Split up cloud testing
